Add support for new Durable Functions backends
Adds new backends for Durable Functions to v3 bundles. In order to
support this, additional changes were made to ensure that the new
extensions were only built in .NET Core 3.1 or greater, as the
extensions rely on .NET Standard 2.1. Therefore, we can not build them
in the v2 bits, so additional metadata was added to reflect this.

Changes will also need to be made to Functions Host to properly load
v3_any bits for the local dev experience.